Update GrassTile.java
I saw it wasn't returning true for both tools; if the stamina is paid and the tile is changed, the interaction should definitely be considered successful. Also, I think it makes more sense that you have to go to dirt before you can go to farmland. Also I did not understand the purpose for the second random check with the hoe, I don't think it should ever drop two seeds.